ref: Moved all cache related custom functions to its own file
