+++ Environment/Libraries/Globals.cpp
@@ -210,16 +210,6 @@ namespace RbxStu {
         return 0;
     }
 
-    int compareinstances(lua_State *L) {
-        Utilities::checkInstance(L, 1, "ANY");
-        Utilities::checkInstance(L, 2, "ANY");
-
-        lua_pushboolean(L, *static_cast<const std::uintptr_t *>(lua_touserdata(L, 1)) ==
-                                   *static_cast<const std::uintptr_t *>(lua_touserdata(L, 2)));
-
-        return 1;
-    }
-
     int fireproximityprompt(lua_State *L) {
         Utilities::checkInstance(L, 1, "ProximityPrompt");
 
@@ -227,30 +217,6 @@ namespace RbxStu {
         reinterpret_cast<RbxStu::StudioFunctionDefinitions::r_RBX_ProximityPrompt_onTriggered>(
                 RobloxManager::GetSingleton()->GetRobloxFunction("RBX::ProximityPrompt::onTriggered"))(proximityPrompt);
         return 0;
-    }
-
-    int cloneref(lua_State *L) {
-        Utilities::checkInstance(L, 1, "ANY");
-
-        const auto userdata = lua_touserdata(L, -1);
-        const auto rawUserdata = *static_cast<void **>(userdata);
-        const auto robloxManager = RobloxManager::GetSingleton();
-        lua_pushlightuserdata(L, robloxManager->GetRobloxFunction("RBX::Instance::pushInstance"));
-        lua_rawget(L, LUA_REGISTRYINDEX);
-
-        lua_pushlightuserdata(L, rawUserdata);
-        lua_rawget(L, -2);
-
-        lua_pushlightuserdata(L, rawUserdata);
-        lua_pushnil(L);
-        lua_rawset(L, -4);
-
-        reinterpret_cast<RbxStu::StudioFunctionDefinitions::r_RBX_Instance_pushInstance>(
-                robloxManager->GetRobloxFunction("RBX::Instance::pushInstance"))(L, userdata);
-        lua_pushlightuserdata(L, rawUserdata);
-        lua_pushvalue(L, -3);
-        lua_rawset(L, -5);
-        return 1;
     }
 
     int lz4compress(lua_State *L) {
@@ -536,9 +502,7 @@ luaL_Reg *Globals::GetLibraryFunctions() {
                                {"isluau", RbxStu::isluau},
                                {"httpget", RbxStu::httpget},
                                {"gethui", RbxStu::gethui},
-                               {"compareinstances", RbxStu::compareinstances},
                                {"fireproximityprompt", RbxStu::fireproximityprompt},
-                               {"cloneref", RbxStu::cloneref},
                                {"lz4compress", RbxStu::lz4compress},
                                {"lz4decompress", RbxStu::lz4decompress},
                                {"messagebox", RbxStu::messagebox},
